@@ -1,33 +1,33 @@
 package com.plusonelabs.calendar;
 
-import java.util.Calendar;
-
 import android.app.AlarmManager;
 import android.app.PendingIntent;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
 
+import java.util.Calendar;
+
 public class AlarmReceiver extends BroadcastReceiver {
 
-	@Override
-	public void onReceive(Context context, Intent intent) {
-		EventAppWidgetProvider.updateAllWidgets(context);
-		EventAppWidgetProvider.updateEventList(context);
-		scheduleAlarm(context);
-	}
+    @Override
+    public void onReceive(Context context, Intent intent) {
+        EventAppWidgetProvider.updateAllWidgets(context);
+        EventAppWidgetProvider.updateEventList(context);
+        scheduleAlarm(context);
+    }
 
-	public static void scheduleAlarm(Context context) {
-		Calendar calendar = Calendar.getInstance();
-		calendar.set(Calendar.HOUR_OF_DAY, 0);
-		calendar.set(Calendar.MINUTE, 0);
-		calendar.set(Calendar.SECOND, 0);
-		calendar.set(Calendar.MILLISECOND, 0);
-		calendar.add(Calendar.DAY_OF_MONTH, 1);
-		PendingIntent pendingIntent = PendingIntent.getBroadcast(context, 0, new Intent(context,
-				AlarmReceiver.class), PendingIntent.FLAG_UPDATE_CURRENT);
-		AlarmManager am = (AlarmManager) context.getSystemService(Context.ALARM_SERVICE);
-		am.set(AlarmManager.RTC, calendar.getTimeInMillis(), pendingIntent);
-	}
+    public static void scheduleAlarm(Context context) {
+        Calendar calendar = Calendar.getInstance();
+        calendar.set(Calendar.HOUR_OF_DAY, 0);
+        calendar.set(Calendar.MINUTE, 0);
+        calendar.set(Calendar.SECOND, 0);
+        calendar.set(Calendar.MILLISECOND, 0);
+        calendar.add(Calendar.DAY_OF_MONTH, 1);
+        PendingIntent pendingIntent = PendingIntent.getBroadcast(context, 0, new Intent(context,
+                AlarmReceiver.class), PendingIntent.FLAG_UPDATE_CURRENT);
+        AlarmManager am = (AlarmManager) context.getSystemService(Context.ALARM_SERVICE);
+        am.set(AlarmManager.RTC, calendar.getTimeInMillis(), pendingIntent);
+    }
 
 }
