@@ -1,22 +1,23 @@
 package com.plusonelabs.calendar;
 
+import android.appwidget.AppWidgetManager;
 import android.content.Context;
 import android.content.Intent;
 import android.view.ContextThemeWrapper;
 import android.widget.RemoteViewsService;
 
-import static com.plusonelabs.calendar.Theme.getCurrentThemeId;
-import static com.plusonelabs.calendar.prefs.CalendarPreferences.PREF_ENTRY_THEME;
-import static com.plusonelabs.calendar.prefs.CalendarPreferences.PREF_ENTRY_THEME_DEFAULT;
+import com.plusonelabs.calendar.prefs.InstanceSettings;
 
+import static com.plusonelabs.calendar.Theme.themeNameToResId;
 
 public class EventWidgetService extends RemoteViewsService {
 
-	@Override
-	public RemoteViewsFactory onGetViewFactory(Intent intent) {
+    @Override
+    public RemoteViewsFactory onGetViewFactory(Intent intent) {
+        int widgetId = intent.getIntExtra(AppWidgetManager.EXTRA_APPWIDGET_ID, 0);
         Context appContext = getApplicationContext();
-        int currentThemeId = getCurrentThemeId(appContext, PREF_ENTRY_THEME, PREF_ENTRY_THEME_DEFAULT);
+        int currentThemeId = themeNameToResId(InstanceSettings.fromId(appContext, widgetId).getEntryTheme());
         ContextThemeWrapper context = new ContextThemeWrapper(appContext, currentThemeId);
-        return new EventRemoteViewsFactory(context);
+        return new EventRemoteViewsFactory(context, widgetId);
     }
 }
