@@ -2,18 +2,18 @@ package com.plusonelabs.calendar;
 
 import android.widget.RemoteViews;
 
-import com.plusonelabs.calendar.model.Event;
+import com.plusonelabs.calendar.widget.WidgetEntry;
 
 import java.util.List;
 
-public interface IEventVisualizer<T extends Event> {
+public interface IEventVisualizer<T extends WidgetEntry> {
 
-	RemoteViews getRemoteView(Event eventEntry);
+    RemoteViews getRemoteView(WidgetEntry eventEntry);
 
-	int getViewTypeCount();
+    int getViewTypeCount();
 
-	List<T> getEventEntries();
+    List<T> getEventEntries();
 
-	Class<? extends T> getSupportedEventEntryType();
+    Class<? extends T> getSupportedEventEntryType();
 
 }
