@@ -1,24 +1,23 @@
 package com.plusonelabs.calendar.calendar;
 
-import android.content.ContentResolver;
 import android.content.ContentUris;
 import android.content.Context;
-import android.content.SharedPreferences;
 import android.database.Cursor;
 import android.net.Uri;
 import android.os.Build;
-import android.preference.PreferenceManager;
 import android.provider.CalendarContract.Attendees;
 import android.provider.CalendarContract.Instances;
-import android.text.format.DateUtils;
+import android.support.annotation.NonNull;
+import android.util.SparseArray;
 
-import com.plusonelabs.calendar.prefs.CalendarPreferences;
+import com.plusonelabs.calendar.DateUtil;
+import com.plusonelabs.calendar.prefs.InstanceSettings;
+import com.plusonelabs.calendar.util.PermissionsUtil;
 
 import org.joda.time.DateTime;
+import org.joda.time.DateTimeZone;
 
 import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Set;
@@ -27,183 +26,243 @@ import static android.graphics.Color.argb;
 import static android.graphics.Color.blue;
 import static android.graphics.Color.green;
 import static android.graphics.Color.red;
-import static com.plusonelabs.calendar.prefs.CalendarPreferences.PREF_EVENT_RANGE;
-import static com.plusonelabs.calendar.prefs.CalendarPreferences.PREF_EVENT_RANGE_DEFAULT;
-import static com.plusonelabs.calendar.prefs.CalendarPreferences.PREF_FILL_ALL_DAY;
-import static com.plusonelabs.calendar.prefs.CalendarPreferences.PREF_FILL_ALL_DAY_DEFAULT;
 
 public class CalendarEventProvider {
 
-	private static final String EVENT_SORT_ORDER = "startDay ASC, allDay DESC, begin ASC ";
-	private static final String EVENT_SELECTION = Instances.SELF_ATTENDEE_STATUS + "!="
-			+ Attendees.ATTENDEE_STATUS_DECLINED;
-    private static final String[] PROJECTION_4_0 = new String[]{Instances.EVENT_ID, Instances.TITLE,
-            Instances.BEGIN, Instances.END, Instances.ALL_DAY, Instances.EVENT_LOCATION,
-            Instances.HAS_ALARM, Instances.RRULE, Instances.CALENDAR_COLOR, Instances.EVENT_COLOR};
-    private static final String[] PROJECTION_4_1 = new String[]{Instances.EVENT_ID, Instances.TITLE,
-            Instances.BEGIN, Instances.END, Instances.ALL_DAY, Instances.EVENT_LOCATION,
-            Instances.HAS_ALARM, Instances.RRULE, Instances.DISPLAY_COLOR};
+    public static final String EVENT_SORT_ORDER = "startDay ASC, allDay DESC, begin ASC ";
+    private static final String EVENT_SELECTION = Instances.SELF_ATTENDEE_STATUS + "!="
+            + Attendees.ATTENDEE_STATUS_DECLINED;
     private static final String CLOSING_BRACKET = " )";
     private static final String OR = " OR ";
-	private static final String EQUALS = " = ";
-	private static final String AND_BRACKET = " AND (";
+    private static final String EQUALS = " = ";
+    private static final String AND_BRACKET = " AND (";
 
-	private final Context context;
+    private final Context context;
+    private final int widgetId;
 
-	public CalendarEventProvider(Context context) {
-		this.context = context;
-	}
+    // Below are parameters, which may change in settings
+    private DateTimeZone zone;
+    private KeywordsFilter mKeywordsFilter;
+    private DateTime mStartOfTimeRange;
+    private DateTime mEndOfTimeRange;
 
-	public List<CalendarEvent> getEvents() {
-		Cursor cursor = createLoadedCursor();
-		if (cursor != null) {
-			List<CalendarEvent> eventList = createEventList(cursor);
-			cursor.close();
-			Collections.sort(eventList);
-			return eventList;
-		}
-        return new ArrayList<>();
+    public CalendarEventProvider(Context context, int widgetId) {
+        this.context = context;
+        this.widgetId = widgetId;
     }
 
-	private List<CalendarEvent> createEventList(Cursor calendarCursor) {
-        SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(context);
-        boolean fillAllDayEvents = prefs.getBoolean(PREF_FILL_ALL_DAY, PREF_FILL_ALL_DAY_DEFAULT);
-        List<CalendarEvent> eventList = new ArrayList<>();
-        for (int i = 0; i < calendarCursor.getCount(); i++) {
-			calendarCursor.moveToPosition(i);
-			CalendarEvent event = createCalendarEvent(calendarCursor);
-			setupDayOneEntry(eventList, event);
-            if (!event.isAllDay() || fillAllDayEvents) {
-                createFollowingEntries(eventList, event);
+    public List<CalendarEvent> getEvents() {
+        initialiseParameters();
+        if (!PermissionsUtil.arePermissionsGranted(context)) {
+            return new ArrayList<>();
+        }
+        List<CalendarEvent> eventList = getTimeFilteredEventList();
+        if (getSettings().getShowPastEventsWithDefaultColor()) {
+            addPastEventsWithDefaultColor(eventList);
+        }
+        if (getSettings().getShowOnlyClosestInstanceOfRecurringEvent()) {
+            filterShowOnlyClosestInstanceOfRecurringEvent(eventList);
+        }
+        return eventList;
+    }
+
+    private void addPastEventsWithDefaultColor(List<CalendarEvent> eventList) {
+        for (CalendarEvent event : getPastEventsWithColorList()) {
+            if (eventList.contains(event)) {
+                eventList.remove(event);
+            }
+            eventList.add(event);
+        }
+    }
+
+    private void filterShowOnlyClosestInstanceOfRecurringEvent(@NonNull List<CalendarEvent> eventList) {
+        SparseArray<CalendarEvent> eventIds = new SparseArray<>();
+        List<CalendarEvent> toDelete = new ArrayList<>();
+        for (CalendarEvent event : eventList) {
+            CalendarEvent otherEvent = eventIds.get(event.getEventId());
+            if (otherEvent == null) {
+                eventIds.put(event.getEventId(), event);
+            } else if (Math.abs(event.getStartDate().getMillis() -
+                    DateUtil.now(zone).getMillis()) <
+                    Math.abs(otherEvent.getStartDate().getMillis() -
+                            DateUtil.now(zone).getMillis())) {
+                toDelete.add(otherEvent);
+                eventIds.put(event.getEventId(), event);
+            } else {
+                toDelete.add(event);
+            }
+        }
+        eventList.removeAll(toDelete);
+    }
+
+    private void initialiseParameters() {
+        zone = getSettings().getTimeZone();
+        mKeywordsFilter = new KeywordsFilter(getSettings().getHideBasedOnKeywords());
+        mStartOfTimeRange = getSettings().getEventsEnded().endedAt(DateUtil.now(zone));
+        mEndOfTimeRange = getEndOfTimeRange(DateUtil.now(zone));
+    }
+
+    public DateTime getEndOfTimeRange() {
+        return mEndOfTimeRange;
+    }
+
+    public DateTime getStartOfTimeRange() {
+        return mStartOfTimeRange;
+    }
+
+    private DateTime getEndOfTimeRange(DateTime now) {
+        int dateRange = getSettings().getEventRange();
+        return dateRange > 0
+                ? now.plusDays(dateRange)
+                : now.withTimeAtStartOfDay().plusDays(1);
+    }
+
+    @NonNull
+    private InstanceSettings getSettings() {
+        return InstanceSettings.fromId(context, widgetId);
+    }
+
+    private List<CalendarEvent> getTimeFilteredEventList() {
+        Uri.Builder builder = Instances.CONTENT_URI.buildUpon();
+        ContentUris.appendId(builder, mStartOfTimeRange.getMillis());
+        ContentUris.appendId(builder, mEndOfTimeRange.getMillis());
+        List<CalendarEvent> eventList = queryList(builder.build(), getCalendarSelection());
+        // Above filters are not exactly correct for AllDay events: for them that filter
+        // time should be moved by a time zone... (i.e. by several hours)
+        // This is why we need to do additional filtering after querying a Content Provider:
+        for (Iterator<CalendarEvent> it = eventList.iterator(); it.hasNext(); ) {
+            CalendarEvent event = it.next();
+            if (!event.getEndDate().isAfter(mStartOfTimeRange)
+                    || !mEndOfTimeRange.isAfter(event.getStartDate())) {
+                // We remove using Iterator to avoid ConcurrentModificationException
+                it.remove();
             }
         }
         return eventList;
     }
 
-	public void setupDayOneEntry(List<CalendarEvent> eventList, CalendarEvent event) {
-		if (isEqualOrAfterTodayAtMidnight(event.getStartDate())) {
-			if (event.daysSpanned() > 1) {
-				CalendarEvent clone = event.clone();
-                clone.setEndDate(event.getStartDate().plusDays(1).withTimeAtStartOfDay());
-                clone.setSpansMultipleDays(true);
-                clone.setOriginalEvent(event);
-                eventList.add(clone);
-            } else {
-				eventList.add(event);
-			}
-		}
-	}
-
-	public void createFollowingEntries(List<CalendarEvent> eventList, CalendarEvent event) {
-		int daysCovered = event.daysSpanned();
-		for (int j = 1; j < daysCovered; j++) {
-            DateTime startDate = event.getStartDate().withTimeAtStartOfDay().plusDays(j);
-            if (isEqualOrAfterTodayAtMidnight(startDate)) {
-                DateTime endDate;
-                if (j < daysCovered - 1) {
-                    endDate = startDate.plusDays(1);
-				} else {
-					endDate = event.getEndDate();
-				}
-				eventList.add(cloneAsSpanningEvent(event, startDate, endDate));
-			}
-		}
-	}
-
-	private boolean isEqualOrAfterTodayAtMidnight(DateTime startDate) {
-        DateTime startOfDay = DateTime.now().withTimeAtStartOfDay();
-        return startDate.isEqual(startOfDay) || startDate.isAfter(startOfDay);
+    private String getCalendarSelection() {
+        Set<String> activeCalendars = getSettings().getActiveCalendars();
+        StringBuilder stringBuilder = new StringBuilder(EVENT_SELECTION);
+        if (!activeCalendars.isEmpty()) {
+            stringBuilder.append(AND_BRACKET);
+            Iterator<String> iterator = activeCalendars.iterator();
+            while (iterator.hasNext()) {
+                String calendarId = iterator.next();
+                stringBuilder.append(Instances.CALENDAR_ID);
+                stringBuilder.append(EQUALS);
+                stringBuilder.append(calendarId);
+                if (iterator.hasNext()) {
+                    stringBuilder.append(OR);
+                }
+            }
+            stringBuilder.append(CLOSING_BRACKET);
+        }
+        return stringBuilder.toString();
     }
 
-    private CalendarEvent cloneAsSpanningEvent(CalendarEvent eventEntry, DateTime startDate,
-                                               DateTime endDate) {
-        CalendarEvent clone = eventEntry.clone();
-		clone.setStartDate(startDate);
-		clone.setEndDate(endDate);
-		clone.setSpansMultipleDays(true);
-		clone.setOriginalEvent(eventEntry);
-		return clone;
-	}
+    private List<CalendarEvent> queryList(Uri uri, String selection) {
+        List<CalendarEvent> eventList = new ArrayList<>();
+        CalendarQueryResult result = new CalendarQueryResult(getSettings(), uri, getProjection(),
+                selection, null, EVENT_SORT_ORDER);
+        Cursor cursor = null;
+        try {
+            cursor = context.getContentResolver().query(uri, getProjection(),
+                    selection, null, EVENT_SORT_ORDER);
+            if (cursor != null) {
+                for (int i = 0; i < cursor.getCount(); i++) {
+                    cursor.moveToPosition(i);
+                    if (CalendarQueryResultsStorage.getNeedToStoreResults()) {
+                        result.addRow(cursor);
+                    }
+                    CalendarEvent event = createCalendarEvent(cursor);
+                    if (!eventList.contains(event) && !mKeywordsFilter.matched(event.getTitle())) {
+                        eventList.add(event);
+                    }
+                }
+            }
+        } finally {
+            if (cursor != null && !cursor.isClosed()) {
+                cursor.close();
+            }
+        }
+        CalendarQueryResultsStorage.store(result);
+        return eventList;
+    }
 
-	private CalendarEvent createCalendarEvent(Cursor calendarCursor) {
-		CalendarEvent event = new CalendarEvent();
-		event.setEventId(calendarCursor.getInt(0));
-		event.setTitle(calendarCursor.getString(1));
-		event.setStartDate(new DateTime(calendarCursor.getLong(2)));
-		event.setEndDate(new DateTime(calendarCursor.getLong(3)));
-		event.setAllDay(calendarCursor.getInt(4) > 0);
-        event.setLocation(calendarCursor.getString(5));
-        event.setAlarmActive(calendarCursor.getInt(6) > 0);
-        event.setRecurring(calendarCursor.getString(7) != null);
-        event.setColor(getAsOpaque(getEventColor(calendarCursor)));
-        if (event.isAllDay()) {
-			DateTime startDate = event.getStartDate();
-			long converted = startDate.getZone().convertLocalToUTC(startDate.getMillis(), true);
-			event.setStartDate(new DateTime(converted));
-			DateTime endDate = event.getEndDate();
-			converted = endDate.getZone().convertLocalToUTC(endDate.getMillis(), true);
-			event.setEndDate(new DateTime(converted));
-		}
-		return event;
-	}
-
-    private int getEventColor(Cursor calendarCursor) {
+    public static String[] getProjection() {
+        List<String> columnNames = new ArrayList<>();
+        columnNames.add(Instances.EVENT_ID);
+        columnNames.add(Instances.TITLE);
+        columnNames.add(Instances.BEGIN);
+        columnNames.add(Instances.END);
+        columnNames.add(Instances.ALL_DAY);
+        columnNames.add(Instances.EVENT_LOCATION);
+        columnNames.add(Instances.HAS_ALARM);
+        columnNames.add(Instances.RRULE);
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) {
-            return calendarCursor.getInt(8);
+            columnNames.add(Instances.DISPLAY_COLOR);
         } else {
-            int eventColor = calendarCursor.getInt(9);
+            columnNames.add(Instances.CALENDAR_COLOR);
+            columnNames.add(Instances.EVENT_COLOR);
+        }
+        return columnNames.toArray(new String[columnNames.size()]);
+    }
+
+    private List<CalendarEvent> getPastEventsWithColorList() {
+        Uri.Builder builder = Instances.CONTENT_URI.buildUpon();
+        ContentUris.appendId(builder, 0);
+        ContentUris.appendId(builder, DateUtil.now(zone).getMillis());
+        List<CalendarEvent> eventList = queryList(builder.build(), getPastEventsWithColorSelection());
+        for (CalendarEvent event : eventList) {
+            event.setDefaultCalendarColor();
+        }
+        return eventList;
+    }
+
+    private String getPastEventsWithColorSelection() {
+        StringBuilder stringBuilder = new StringBuilder(getCalendarSelection());
+        stringBuilder.append(AND_BRACKET);
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) {
+            stringBuilder.append(Instances.DISPLAY_COLOR);
+            stringBuilder.append(EQUALS);
+            stringBuilder.append(Instances.CALENDAR_COLOR);
+        } else {
+            stringBuilder.append(Instances.EVENT_COLOR);
+            stringBuilder.append(EQUALS);
+            stringBuilder.append("0");
+        }
+        stringBuilder.append(CLOSING_BRACKET);
+        return stringBuilder.toString();
+    }
+
+    private CalendarEvent createCalendarEvent(Cursor cursor) {
+        boolean allDay = cursor.getInt(cursor.getColumnIndex(Instances.ALL_DAY)) > 0;
+        CalendarEvent event = new CalendarEvent(context, widgetId, zone, allDay);
+        event.setEventId(cursor.getInt(cursor.getColumnIndex(Instances.EVENT_ID)));
+        event.setTitle(cursor.getString(cursor.getColumnIndex(Instances.TITLE)));
+        event.setStartMillis(cursor.getLong(cursor.getColumnIndex(Instances.BEGIN)));
+        event.setEndMillis(cursor.getLong(cursor.getColumnIndex(Instances.END)));
+        event.setLocation(cursor.getString(cursor.getColumnIndex(Instances.EVENT_LOCATION)));
+        event.setAlarmActive(cursor.getInt(cursor.getColumnIndex(Instances.HAS_ALARM)) > 0);
+        event.setRecurring(cursor.getString(cursor.getColumnIndex(Instances.RRULE)) != null);
+        event.setColor(getAsOpaque(getEventColor(cursor)));
+        return event;
+    }
+
+    private int getEventColor(Cursor cursor) {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) {
+            return cursor.getInt(cursor.getColumnIndex(Instances.DISPLAY_COLOR));
+        } else {
+            int eventColor = cursor.getInt(cursor.getColumnIndex(Instances.EVENT_COLOR));
             if (eventColor > 0) {
                 return eventColor;
             }
-            return calendarCursor.getInt(8);
+            return cursor.getInt(cursor.getColumnIndex(Instances.CALENDAR_COLOR));
         }
     }
 
     private int getAsOpaque(int color) {
         return argb(255, red(color), green(color), blue(color));
-	}
-
-	private Cursor createLoadedCursor() {
-		SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(context);
-		int dateRange = Integer
-				.valueOf(prefs.getString(PREF_EVENT_RANGE, PREF_EVENT_RANGE_DEFAULT));
-		long start = System.currentTimeMillis();
-		long end = start + DateUtils.DAY_IN_MILLIS * dateRange;
-		Uri.Builder builder = Instances.CONTENT_URI.buildUpon();
-		ContentUris.appendId(builder, start);
-		ContentUris.appendId(builder, end);
-		String selection = createSelectionClause();
-		ContentResolver contentResolver = context.getContentResolver();
-        return contentResolver.query(builder.build(), getProjection(), selection, null, EVENT_SORT_ORDER);
-    }
-
-    private String[] getProjection() {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) {
-            return PROJECTION_4_1;
-        }
-        return PROJECTION_4_0;
-    }
-
-    private String createSelectionClause() {
-		SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(context);
-		Set<String> activeCalenders = prefs.getStringSet(CalendarPreferences.PREF_ACTIVE_CALENDARS,
-				new HashSet<String>());
-		if (activeCalenders.isEmpty()) {
-			return EVENT_SELECTION;
-		}
-        StringBuilder stringBuilder = new StringBuilder();
-        stringBuilder.append(AND_BRACKET);
-        Iterator<String> iterator = activeCalenders.iterator();
-		while (iterator.hasNext()) {
-			String calendarId = iterator.next();
-            stringBuilder.append(Instances.CALENDAR_ID);
-            stringBuilder.append(EQUALS);
-            stringBuilder.append(calendarId);
-            if (iterator.hasNext()) {
-                stringBuilder.append(OR);
-            }
-		}
-        stringBuilder.append(CLOSING_BRACKET);
-        return EVENT_SELECTION + stringBuilder.toString();
     }
 }
