@@ -8,16 +8,16 @@ import com.plusonelabs.calendar.R;
 
 public class EventDetailsPreferencesFragment extends PreferenceFragment {
 
-	@Override
-	public void onCreate(Bundle savedInstanceState) {
-		super.onCreate(savedInstanceState);
-		addPreferencesFromResource(R.xml.preferences_event_details);
-	}
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        addPreferencesFromResource(R.xml.preferences_event_details);
+    }
 
-	@Override
-	public void onPause() {
-		super.onPause();
-		EventAppWidgetProvider.updateEventList(getActivity());
-		EventAppWidgetProvider.updateAllWidgets(getActivity());
-	}
+    @Override
+    public void onPause() {
+        super.onPause();
+        EventAppWidgetProvider.updateEventList(getActivity());
+        EventAppWidgetProvider.updateAllWidgets(getActivity());
+    }
 }
